Moved tests into test folder to fix failed netlify deploy builds. Added @babel/plugin-proposal-private-property-in-object to dev dependencies to eliminate an error on netlify build
